+++ packages/activity_recognition_flutter/example/android/app/src/main/java/dk/cachet/activity_recognition_flutter_example/MainActivity.java
@@ -1,0 +1,6 @@
+package dk.cachet.activity_recognition_flutter_example;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}