+++ packages/health/example/android/app/src/main/kotlin/cachet/plugins/health_example/MainActivity.kt
@@ -1,12 +1,8 @@
 package cachet.plugins.health_example
 
-import androidx.annotation.NonNull
-import io.flutter.embedding.engine.FlutterEngine
+import android.os.Bundle
+
 import io.flutter.embedding.android.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
 
 class MainActivity: FlutterActivity() {
-  override fun configureFlutterEngine(@NonNull flutterEngine: FlutterEngine) {
-    GeneratedPluginRegistrant.registerWith(flutterEngine)
-  }
 }