+++ packages/health/example/ios/Flutter/flutter_export_environment.sh
@@ -1,7 +1,7 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/home/ikeyan/development/flutter"
-export "FLUTTER_APPLICATION_PATH=/home/ikeyan/work/forks/flutter-plugins/packages/health/example"
+export "FLUTTER_ROOT=/Users/bardram/dev/flutter"
+export "FLUTTER_APPLICATION_PATH=/Users/bardram/dev/flutter-plugins/packages/health/example"
 export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"