+++ packages/health/ios/Classes/SwiftHealthPlugin.swift
@@ -123,6 +123,8 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
                         "value": sample.value,
                         "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
                         "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
+                        "source_id": sample.sourceRevision.source.bundleIdentifier,
+                        "source_name": sample.sourceRevision.source.name
                     ]
                 })
                 return
@@ -135,6 +137,8 @@ public class SwiftHealthPlugin: NSObject, FlutterPlugin {
                     "value": sample.quantity.doubleValue(for: unit),
                     "date_from": Int(sample.startDate.timeIntervalSince1970 * 1000),
                     "date_to": Int(sample.endDate.timeIntervalSince1970 * 1000),
+                    "source_id": sample.sourceRevision.source.bundleIdentifier,
+                    "source_name": sample.sourceRevision.source.name
                 ]
             })
             return