+++ packages/mobility_features/test/mobility_features_test.dart
@@ -19,9 +19,11 @@ void main() async {
   GeoLocation pos2 = GeoLocation(55.685329, 12.538601);
 
   // Dronning Louises Bro
+  // ignore: unused_local_variable
   GeoLocation pos3 = GeoLocation(55.686723, 12.563769);
 
   // Assistentens Kirkegaard
+  // ignore: unused_local_variable
   GeoLocation pos4 = GeoLocation(55.690862, 12.549545);
 
   /// This test  verifies that the 'midnight' extension
@@ -417,7 +419,6 @@ void main() async {
       Stream<MobilityContext> mobilityStream = factory.contextStream;
 
       mobilityStream.listen(expectAsync1((event) {
-
         print('=' * 150);
         print(
             "Mobility Context Received: ${jsonEncoder.convert(event.toJson())}");